llm/requirements.txt
transformers>=4.40.0
datasets>=2.18.0
accelerate>=0.27.0
huggingface-hub>=0.20.0
llama-cpp-python>=0.2.56
psutil>=5.9.0
PyGithub>=2.1.1
PyYAML>=6.0.1
requests>=2.31.0
inquirer>=3.1.3
streamlit>=1.29.0
plotly>=5.18.0
pandas>=2.2.0
wordcloud>=1.9.3
networkx>=3.2.1
matplotlib>=3.8.2
click>=8.1.7
rich>=13.7.0
tensorboard>=2.15.1
scikit-learn>=1.4.0
wandb>=0.16.2
shap>=0.44.0
lime>=0.2.0.1
seaborn>=0.13.2
numpy>=1.26.3
